+++ LuckParser/Models/ParseModels/Mechanics/PlayerCastEndMechanic.cs
@@ -1,0 +1,48 @@
+﻿using LuckParser.Controllers;
+using LuckParser.Parser;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace LuckParser.Models.ParseModels
+{
+    
+    public class PlayerCastEndMechanic : Mechanic
+    {
+
+        public PlayerCastEndMechanic(long skillId, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, int internalCoolDown, List<MechanicChecker> conditions, TriggerRule rule) : this(skillId, inGameName, plotlySetting, shortName, shortName, shortName, internalCoolDown, conditions, rule)
+        {
+        }
+
+        public PlayerCastEndMechanic(long skillId, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown) : base(skillId, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown)
+        {
+        }
+
+        public PlayerCastEndMechanic(long skillId, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, int internalCoolDown) : this(skillId, inGameName, plotlySetting, shortName, shortName, shortName, internalCoolDown)
+        {
+        }
+
+        public PlayerCastEndMechanic(long skillId, string inGameName, MechanicPlotlySetting plotlySetting, string shortName, string description, string fullName, int internalCoolDown, List<MechanicChecker> conditions, TriggerRule rule) : base(skillId, inGameName, plotlySetting, shortName, description, fullName, internalCoolDown, conditions, rule)
+        {
+        }
+
+        public override void CheckMechanic(ParsedLog log, Dictionary<ushort, DummyActor> regroupedMobs)
+        {
+            MechanicData mechData = log.MechanicData;
+            CombatData combatData = log.CombatData;
+            HashSet<ushort> playersIds = log.PlayerIDs;
+            foreach (Player p in log.PlayerList)
+            {
+                foreach (CombatItem c in log.CombatData.GetCastDataById(SkillId))
+                {
+                    if (!c.IsActivation.StartCasting() && c.SrcInstid == p.InstID && Keep(c, log))
+                    {
+                        mechData[this].Add(new MechanicLog(log.FightData.ToFightSpace(c.Time), this, p));
+
+                    }
+                }
+            }
+        }
+    }
+}
